Move last Views into GUI package and set as Deprecated for removal
