+++ src/GUI/package-info.java
@@ -1,0 +1,2 @@
+@Deprecated(forRemoval = true)
+package GUI;
